24ms 92.31%, 11.7MB 21.05%.
***** NOTICE *****
From this time, we will not upload the whole VS project, neither we will not use any IDE to finish the problem.

We will just upload the segment of code where we can finish it on LeetCode browser coding platform.

That is, you can just paste the code on the platform and see the result :)
*****